Updated: util.locations File(s) names to proper java naming conventions. Mining.java Woodcutting.java WoodcuttingPoly.java
Added:
[LocalPlayer.java]
#isInGE - Checks to see if local character is at the Grand Exchange.
#walkToGE - Walks your player to the Grand Exchange.
#openGE - Opens the Grand Exchange. (Handles checking if you're in GE + Walking there if not).

[GrandExchange.java]
#PurchaseItem.class
Custom class that handles extracting data from an ArrayList into ItemID, quantity, & multipliedValue

#buy - ArrayList support
#buy - Single item support

Handles updating the Grand Exchange prices every 30 minutes, if it's not already updated.
Checks to see if your total coin value, has been checked. If not, it will check your bank and inventory, and set the total amount to an int.
Checks to see if your desired buy price is higher than your total coins value.
If it's higher, it will check to see if you have enough coins to buy the item without any multiplied value.
If you can buy with a multiplied value, it will set the value to the max it can (if your supplied multiplied value is higher than your coin value you can afford)
If you can't buy the item with any multiplied value, at the price of the item, it will fail the purchase, setting a boolean to true for later use.
If you can buy the item, and the GE is not open, it will handle walking there, and opening the GE.
Finally, it will buy the item(s) at the desired quantity & price (multipliedValue).

util.locations
[Constants.java]
#GRAND_EXCHANGE